JD-117 add QueryDSL support & use single query for getting registry error counts

+++ src/main/java/ru/mail/jira/plugins/groovy/api/dto/directory/RegistryScriptDto.java
@@ -23,5 +23,5 @@ public class RegistryScriptDto extends RegistryScriptForm {
     @XmlElement
     private List<ScriptParamDto> params;
     @XmlElement
-    private Integer errorCount;
+    private Long errorCount;
 }
